Add patch from upstream to make the setsid() that addresses CVE-2017-5226 optional
It breaks interactive shells. Users of bubblewrap to confine untrusted
programs should either add --new-session to the bwrap command line, or
prevent the TIOCSTI ioctl with a seccomp filter instead (as Flatpak does).

d/control: add Breaks on versions of Flatpak that did not
load the necessary seccomp filter to prevent CVE-2017-5226.